Cleaned up more casts and associated code [PCL-4230]

+++ common/include/pcl/PCLPointField.h
@@ -50,7 +50,7 @@ namespace pcl
       case ::pcl::PCLPointField::PointFieldTypes::UINT32: s << "  UINT32" << std::endl; break;
       case ::pcl::PCLPointField::PointFieldTypes::FLOAT32: s << "  FLOAT32" << std::endl; break;
       case ::pcl::PCLPointField::PointFieldTypes::FLOAT64: s << "  FLOAT64" << std::endl; break;
-      default: s << "  " << static_cast<int>(v.datatype) << std::endl;
+      default: s << "  " << v.datatype << std::endl;
     }
     s << "count: ";
     s << "  " << v.count << std::endl;
